+++ runner/options.go
@@ -247,6 +247,7 @@ type Options struct {
 	OutputMatchResponseTime   string
 	OutputFilterResponseTime  string
 	HealthCheck               bool
+	ListDSL                   bool
 }
 
 // ParseOptions parses the command line options for application
@@ -329,6 +330,7 @@ func ParseOptions() *Options {
 		flagSet.BoolVar(&options.Pipeline, "pipeline", false, "probe and display server supporting HTTP1.1 pipeline"),
 		flagSet.BoolVar(&options.HTTP2Probe, "http2", false, "probe and display server supporting HTTP2"),
 		flagSet.BoolVar(&options.VHost, "vhost", false, "probe and display server supporting VHOST"),
+		flagSet.BoolVar(&options.ListDSL, "list-dsl", false, "list json output field keys name that support dsl matcher/filter"),
 	)
 
 	flagSet.CreateGroup("output", "Output",
@@ -406,7 +408,12 @@ func ParseOptions() *Options {
 	if err != nil {
 		gologger.Fatal().Msgf("%s\n", err)
 	}
-
+	if options.ListDSL {
+		for _, dsl := range dslCommands {
+			gologger.Print().Msg(dsl)
+		}
+		os.Exit(0)
+	}
 	showBanner()
 
 	if options.Version {
